fix: flow change with public domain guard
