+++ src/pages/workspace/travel/PolicyTravelPage.tsx
@@ -19,7 +19,6 @@ import Navigation from '@libs/Navigation/Navigation';
 import type {PlatformStackScreenProps} from '@libs/Navigation/PlatformStackNavigation/types';
 import type {WorkspaceSplitNavigatorParamList} from '@libs/Navigation/types';
 import {getTravelStep} from '@libs/PolicyUtils';
-import {getTravelInvoicingCardSettingsKey, hasTravelInvoicingSettlementAccount} from '@libs/TravelInvoicingUtils';
 import AccessOrNotFoundWrapper from '@pages/workspace/AccessOrNotFoundWrapper';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
@@ -49,10 +48,6 @@ function WorkspaceTravelPage({
     const {login: currentUserLogin} = useCurrentUserPersonalDetails();
     const [policies] = useOnyx(ONYXKEYS.COLLECTION.POLICY, {canBeMissing: false});
 
-    // Get Travel Invoicing card settings to check for settlement account
-    const [cardSettings] = useOnyx(getTravelInvoicingCardSettingsKey(workspaceAccountID), {canBeMissing: true});
-    const hasSettlementAccount = hasTravelInvoicingSettlementAccount(cardSettings);
-
     const fetchTravelData = useCallback(() => {
         openPolicyTravelPage(policyID, workspaceAccountID);
     }, [policyID, workspaceAccountID]);
@@ -79,9 +74,6 @@ function WorkspaceTravelPage({
     const mainContent = (() => {
         // TODO: Remove this conditional when Travel Invoicing feature is fully implemented
         if (isTravelInvoicingEnabled) {
-            if (!hasSettlementAccount) {
-                return <GetStartedTravel policyID={policyID} />;
-            }
             return <WorkspaceTravelInvoicingSection policyID={policyID} />;
         }
         switch (step) {
