+++ src/pages/workspace/travel/WorkspaceTravelInvoicingSection.tsx
@@ -1,4 +1,4 @@
-import React, {useState} from 'react';
+import React from 'react';
 import {View} from 'react-native';
 import AnimatedSubmitButton from '@components/AnimatedSubmitButton';
 import MenuItem from '@components/MenuItem';
@@ -12,12 +12,13 @@ import useSingleExecution from '@hooks/useSingleExecution';
 import useThemeStyles from '@hooks/useThemeStyles';
 import useWorkspaceAccountID from '@hooks/useWorkspaceAccountID';
 import {openExternalLink} from '@libs/actions/Link';
-import {clearTravelInvoicingSettlementAccountErrors} from '@libs/actions/TravelInvoicing';
+import {clearTravelInvoicingSettlementAccountErrors, setTravelInvoicingSettlementAccount} from '@libs/actions/TravelInvoicing';
 import {getLastFourDigits} from '@libs/BankAccountUtils';
+import {getEligibleBankAccountsForCard} from '@libs/CardUtils';
 import {convertToDisplayString} from '@libs/CurrencyUtils';
 import Navigation from '@libs/Navigation/Navigation';
+import {hasInProgressUSDVBBA, REIMBURSEMENT_ACCOUNT_ROUTE_NAMES} from '@libs/ReimbursementAccountUtils';
 import {
-    getIsTravelInvoicingEnabled,
     getTravelInvoicingCardSettingsKey,
     getTravelLimit,
     getTravelSettlementAccount,
@@ -30,7 +31,6 @@ import type {ToggleSettingOptionRowProps} from '@pages/workspace/workflows/Toggl
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import ROUTES from '@src/ROUTES';
-import BookOrManageYourTrip from './BookOrManageYourTrip';
 import CentralInvoicingLearnHow from './CentralInvoicingLearnHow';
 import CentralInvoicingSubtitleWrapper from './CentralInvoicingSubtitleWrapper';
 
@@ -50,16 +50,15 @@ function WorkspaceTravelInvoicingSection({policyID}: WorkspaceTravelInvoicingSec
     const {isExecuting, singleExecution} = useSingleExecution();
     const icons = useMemoizedLazyExpensifyIcons(['LuggageWithLines', 'NewWindow']);
 
-    const [isCentralInvoicingEnabled, setIsCentralInvoicingEnabled] = useState(true);
-
     // For Travel Invoicing, we use a travel-specific card settings key
     // The format is: private_expensifyCardSettings_{workspaceAccountID}_{feedType}
     // where feedType is PROGRAM_TRAVEL_US for Travel Invoicing
     const [cardSettings] = useOnyx(getTravelInvoicingCardSettingsKey(workspaceAccountID), {canBeMissing: true});
+    const [account] = useOnyx(ONYXKEYS.ACCOUNT, {canBeMissing: true});
     const [bankAccountList] = useOnyx(ONYXKEYS.BANK_ACCOUNT_LIST, {canBeMissing: true});
+    const [reimbursementAccount] = useOnyx(ONYXKEYS.REIMBURSEMENT_ACCOUNT, {canBeMissing: true});
 
     // Use pure selectors to derive state
-    const isTravelInvoicingEnabled = getIsTravelInvoicingEnabled(cardSettings);
     const hasSettlementAccount = hasTravelInvoicingSettlementAccount(cardSettings);
     const travelSpend = getTravelSpend(cardSettings);
     const travelLimit = getTravelLimit(cardSettings);
@@ -79,8 +78,41 @@ function WorkspaceTravelInvoicingSection({policyID}: WorkspaceTravelInvoicingSec
     // Get any errors from the settlement account update
     const hasSettlementAccountError = Object.keys(cardSettings?.errors ?? {}).length > 0;
 
+    // Bank account eligibility for toggle handler
+    const isSetupUnfinished = hasInProgressUSDVBBA(reimbursementAccount?.achData);
+    const eligibleBankAccounts = getEligibleBankAccountsForCard(bankAccountList);
+
+    /**
+     * Handle toggle change for Central Invoicing.
+     * When turning ON, triggers bank account flow if needed.
+     * When turning OFF, unassigns the settlement account.
+     */
+    const handleToggle = (isEnabled: boolean) => {
+        if (!isEnabled) {
+            // Turning OFF - unassign the settlement account by setting paymentBankAccountID to 0
+            setTravelInvoicingSettlementAccount(policyID, workspaceAccountID, 0, cardSettings?.paymentBankAccountID);
+            return;
+        }
+
+        // Check if user is on a public domain - Travel Invoicing requires a private domain
+        if (account?.isFromPublicDomain) {
+            Navigation.navigate(ROUTES.TRAVEL_PUBLIC_DOMAIN_ERROR.getRoute(Navigation.getActiveRoute()));
+            return;
+        }
+
+        // Turning ON - check if bank account setup is needed
+        if (!eligibleBankAccounts.length || isSetupUnfinished) {
+            // No bank accounts - start add bank account flow
+            Navigation.navigate(ROUTES.BANK_ACCOUNT_WITH_STEP_TO_OPEN.getRoute(policyID, REIMBURSEMENT_ACCOUNT_ROUTE_NAMES.NEW, ROUTES.WORKSPACE_TRAVEL.getRoute(policyID)));
+            return;
+        }
+
+        // Bank accounts exist - go to settlement account selection
+        Navigation.navigate(ROUTES.WORKSPACE_TRAVEL_SETTINGS_ACCOUNT.getRoute(policyID));
+    };
+
     const getCentralInvoicingSubtitle = () => {
-        if (!isCentralInvoicingEnabled) {
+        if (!hasSettlementAccount) {
             return <CentralInvoicingSubtitleWrapper htmlComponent={<CentralInvoicingLearnHow />} />;
         }
         return <CentralInvoicingSubtitleWrapper />;
@@ -91,8 +123,8 @@ function WorkspaceTravelInvoicingSection({policyID}: WorkspaceTravelInvoicingSec
             title: translate('workspace.moreFeatures.travel.travelInvoicing.centralInvoicingSection.title'),
             subtitle: getCentralInvoicingSubtitle(),
             switchAccessibilityLabel: translate('workspace.moreFeatures.travel.travelInvoicing.centralInvoicingSection.subtitle'),
-            isActive: isCentralInvoicingEnabled,
-            onToggle: (isEnabled: boolean) => setIsCentralInvoicingEnabled(isEnabled),
+            isActive: hasSettlementAccount,
+            onToggle: handleToggle,
             // pendingAction: policy?.pendingFields?.autoReporting ?? policy?.pendingFields?.autoReportingFrequency,
             // errors: getLatestErrorField(policy ?? {}, CONST.POLICY.COLLECTION_KEYS.AUTOREPORTING),
             // onCloseError: () => clearPolicyErrorField(route.params.policyID, CONST.POLICY.COLLECTION_KEYS.AUTOREPORTING),
@@ -184,11 +216,6 @@ function WorkspaceTravelInvoicingSection({policyID}: WorkspaceTravelInvoicingSec
             />
         </Section>
     );
-
-    // If Travel Invoicing beta is not enabled, show the BookOrManageYourTrip component as fallback (before Travel Invoicing feature)
-    if (!isTravelInvoicingEnabled) {
-        return <BookOrManageYourTrip policyID={policyID} />;
-    }
 
     return (
         <>
